se agrega funcionalida de la app

+++ app.js
@@ -1,23 +1,25 @@
 //Selectors***********************************
-const listElement = document.getElementById("main-list");
+const mainListElement = document.getElementById("main-list");
+const secondaryListElement = document.getElementById("secondary-list");
 
 //EventsListener***********************************
-document.addEventListener("DOMContentLoaded", initialLoad);
-
+document.addEventListener("DOMContentLoaded", validateLocalData);
+mainListElement.addEventListener("click", checkUncheck);
+secondaryListElement.addEventListener("click",checkUncheck);
 //Variables***********************************
-
-
+let dataInLocalStorage = localStorage.getItem("MISSIONES-MASS-EFFECT");
+let missionsList =
+  dataInLocalStorage == false ? [] : JSON.parse(dataInLocalStorage);
 //funciones***********************************
 
 // initial load
-function initialLoad() {
-  const dataInLocalStorage = localStorage.getItem("MISSIONES-MASS-EFFECT");
-  if (dataInLocalStorage) {
-    // console.log(JSON.parse(dataInLocalStorage));
-    const missions = JSON.parse(dataInLocalStorage);
-    loadMissions(missions);
+function validateLocalData() {
+  if (missionsList) {
+    displayMissions();
   } else {
     getJson();
+    location.reload;
+    displayMissions();
   }
 }
 
@@ -26,31 +28,109 @@ function getJson() {
   fetch("./misiones.json")
     .then((response) => response.json())
     .then((data) => {
-      const missionsList = data;
-      updateLocalStorage(missionsList);
-      initialLoad();
+      missionsList = data;
+      updateLocalStorage();
     })
     .catch((error) => console.log(error));
 }
 
-// load data and display it
-function loadMissions(data) {
-  const mainMissions = data.principales;
-  const typeInsertAdjacent = "beforeend";
+// display missions
+function displayMissions() {
+  mainListElement.innerHTML = "";
+  secondaryListElement.innerHTML = "";
 
+  const mainMissions = missionsList.principales;
+  const secondaryMission = missionsList.secundarias;
+
+  const typeInsertAdjacent = "beforeend";
+  let idProperty = 0;
+
+  //create elements to main mission
   mainMissions.forEach((element) => {
-    const liHtml = `<li><i class="fas fa-star"></i>${element}</li>`;
-    listElement.insertAdjacentHTML(typeInsertAdjacent, liHtml);
+    const missionStatus = missionsList.principales[idProperty].status
+      ? "completed"
+      : "";
+
+    const iconStatusMain = missionsList.principales[idProperty].status
+      ? '<i class="fas fa-check"></i>'
+      : '<i class="fas fa-star"></i>';
+
+    const liHtml = `<li id="${idProperty}" class=${missionStatus}>${iconStatusMain} ${element.titulo}</li>`;
+
+    mainListElement.insertAdjacentHTML(typeInsertAdjacent, liHtml);
+
+    idProperty++;
+  });
+
+  //create elements to secondary missions
+  secondaryMission.forEach((element, index) => {
+    
+    const newLi = document.createElement("li");
+    const newH4 = document.createElement("h4");
+    const newUl = document.createElement("ul");
+
+    newLi.appendChild(newH4);
+    newLi.appendChild(newUl);
+    newLi.setAttribute("id",index)
+
+    newH4.innerText = element.titulo;
+
+    secondaryListElement.appendChild(newLi);
+
+    const missions = element.misiones;
+
+    missions.forEach((element, index) => {   
+      const newLi = document.createElement("li");
+      
+      newUl.appendChild(newLi);
+      newLi.setAttribute("id",index);
+
+      if(missionStatus = element.status){
+        newLi.classList.add("completed");
+      }
+
+      const iconStatusSecondary = element.status
+        ? '<i class="fas fa-check"></i>'
+        : '<i class="fas fa-circle"></i>';
+
+      newLi.innerHTML = `${iconStatusSecondary}${element.titulo}`;
+      
+    });
   });
 }
 
 // updata the data in local storage
-function updateLocalStorage(missions) {
-  localStorage.setItem("MISSIONES-MASS-EFFECT", JSON.stringify(missions));
+function updateLocalStorage() {
+  localStorage.setItem("MISSIONES-MASS-EFFECT", JSON.stringify(missionsList));
 }
 
 // reset local storage
 function resetLocal() {
+  if (missionsList) {
     localStorage.clear();
     location.reload();
+  }
+}
+//completed
+function checkUncheck(e) {
+  const idMission = e.target.id;
+  const idSecondary = e.target.parentElement.parentElement.id
+
+  const statusMainMission = missionsList.principales[idMission].status;
+  const statusSecondaryMission = missionsList.secundarias[idSecondary].misiones[idMission].status;
+
+  if (statusMainMission) {
+    missionsList.principales[idMission].status = false;
+  } else {
+    missionsList.principales[idMission].status = true;
+  }
+
+  if(statusSecondaryMission){
+    missionsList.secundarias[idSecondary].misiones[idMission].status = false;
+  }else{
+    missionsList.secundarias[idSecondary].misiones[idMission].status = true;
+  }
+
+  updateLocalStorage(missionsList);
+  displayMissions();
 }
